requirements.txt
pyacoustid
eyed3
pillow